requirements.txt
Pygments==2.5.2
argparse==1.2.1
flake8==2.1.0
mccabe==0.2.1
mock==2.0.0
pynose==1.5.2
pep8==1.4.6
pudb==2013.5.1
pyflakes==0.7.3
urwid==1.1.2
wsgiref==0.1.2
